Use a BaseWidget class
